# lib/api-zod/src/generated/types/adminEntityInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * School Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminEntityInputMetadata } from './adminEntityInputMetadata';

export interface AdminEntityInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  parentId?: number | null;
  metadata?: AdminEntityInputMetadata;
}