# lib/api-zod/src/generated/types/board.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * School Learning Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { EntityRef } from './entityRef';

export type Board = EntityRef;