# lib/api-zod/src/generated/types/chapter.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * School Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Chapter {
  id: number;
  name: string;
  bookId: number;
  orderIndex: number;
}